+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 
 - Add "page-break-before" class on import to 3 subsections:
   - "eligibility", "program description", and "application checklist"
+- Add "designer" field to Nofo object
 
 ### Changed